requirements.txt
pandas
pyscbwrapper
requests
psycopg2
sqlalchemy==1.4.42
google-cloud-pubsub
